move standalone tcp to proper package
