+++ akka-http-core/src/main/scala/akka/http/impl/engine/server/NewTcpRegistry.scala
@@ -1,15 +1,15 @@
-package akka.io
+package akka.http.impl.engine.server
 
 import akka.actor.ExtendedActorSystem
 import akka.actor.Extension
 import akka.actor.ExtensionId
 import akka.actor.ExtensionIdProvider
 import akka.dispatch.ExecutionContexts.sameThreadExecutionContext
-import akka.io.SelectionHandler.ChannelRegistryImpl
+import SelectionHandler.ChannelRegistryImpl
 import akka.util.SerializedSuspendableExecutionContext
 
 class ChannelRegistryExt(implicit val system: ExtendedActorSystem) extends akka.actor.Extension {
-  private val settings = Tcp(system).Settings
+  private val settings = akka.io.Tcp(system).Settings
   private val SelectorDispatcher = settings.SelectorDispatcher
   val registry = {
     val dispatcher = system.dispatchers.lookup(SelectorDispatcher)
